stream: move prefixed files into internal/streams.
This simplifies the lib/ folder, and hides _stream_readable.js and
peers into lib/internal/streams. All the tests have been updated to
use the main stream module, rather than requiring from the internals.
Access to the previous files has been deprecated, and it will be removed
in a future release.
