+++ lib/internal/streams/passthrough.js
@@ -19,20 +19,25 @@
 // OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
 // USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+// a passthrough stream.
+// basically just the most minimal sort of Transform stream.
+// Every written chunk gets output as-is.
+
 'use strict';
-const common = require('../common');
 
-const Readable = require('stream').Readable;
-const EE = require('events').EventEmitter;
+module.exports = PassThrough;
 
-const oldStream = new EE();
-oldStream.pause = common.noop;
-oldStream.resume = common.noop;
+const Transform = require('internal/streams/transform');
+const util = require('util');
+util.inherits(PassThrough, Transform);
 
-const newStream = new Readable().wrap(oldStream);
+function PassThrough(options) {
+  if (!(this instanceof PassThrough))
+    return new PassThrough(options);
 
-newStream
-  .on('readable', common.noop)
-  .on('end', common.mustCall());
+  Transform.call(this, options);
+}
 
-oldStream.emit('end');
+PassThrough.prototype._transform = function(chunk, encoding, cb) {
+  cb(null, chunk);
+};
