+++ lib/zlib.js
@@ -23,7 +23,7 @@
 
 const Buffer = require('buffer').Buffer;
 const internalUtil = require('internal/util');
-const Transform = require('_stream_transform');
+const Transform = require('internal/streams/transform');
 const binding = process.binding('zlib');
 const assert = require('assert').ok;
 const kMaxLength = require('buffer').kMaxLength;
