+++ test/message/stdin_messages.out
@@ -11,7 +11,7 @@ SyntaxError: Strict mode code may not include a with statement
     at Socket.<anonymous> (bootstrap_node.js:*:*)
     at emitNone (events.js:*:*)
     at Socket.emit (events.js:*:*)
-    at endReadableNT (_stream_readable.js:*:*)
+    at endReadableNT (internal/streams/readable.js:*:*)
     at _combinedTickCallback (internal/process/next_tick.js:*:*)
 42
 42
@@ -29,7 +29,7 @@ Error: hello
     at Socket.<anonymous> (bootstrap_node.js:*:*)
     at emitNone (events.js:*:*)
     at Socket.emit (events.js:*:*)
-    at endReadableNT (_stream_readable.js:*:*)
+    at endReadableNT (internal/streams/readable.js:*:*)
 [stdin]:1
 throw new Error("hello")
 ^
@@ -44,7 +44,7 @@ Error: hello
     at Socket.<anonymous> (bootstrap_node.js:*:*)
     at emitNone (events.js:*:*)
     at Socket.emit (events.js:*:*)
-    at endReadableNT (_stream_readable.js:*:*)
+    at endReadableNT (internal/streams/readable.js:*:*)
 100
 [stdin]:1
 var x = 100; y = x;
@@ -60,7 +60,7 @@ ReferenceError: y is not defined
     at Socket.<anonymous> (bootstrap_node.js:*:*)
     at emitNone (events.js:*:*)
     at Socket.emit (events.js:*:*)
-    at endReadableNT (_stream_readable.js:*:*)
+    at endReadableNT (internal/streams/readable.js:*:*)
 
 [stdin]:1
 var ______________________________________________; throw 10
