+++ test/parallel/test-stream-pipe-after-end.js
@@ -22,8 +22,8 @@
 'use strict';
 const common = require('../common');
 const assert = require('assert');
-const Readable = require('_stream_readable');
-const Writable = require('_stream_writable');
+const Readable = require('stream').Readable;
+const Writable = require('stream').Writable;
 const util = require('util');
 
 util.inherits(TestReadable, Readable);
