+++ test/parallel/test-stream2-compatibility.js
@@ -21,8 +21,8 @@
 
 'use strict';
 require('../common');
-const R = require('_stream_readable');
-const W = require('_stream_writable');
+const R = require('stream').Readable;
+const W = require('stream').Writable;
 const assert = require('assert');
 
 const util = require('util');
