+++ test/parallel/test-stream2-readable-wrap.js
@@ -22,8 +22,8 @@
 'use strict';
 const common = require('../common');
 const assert = require('assert');
-const Readable = require('_stream_readable');
-const Writable = require('_stream_writable');
+const Readable = require('stream').Readable;
+const Writable = require('stream').Writable;
 const EE = require('events').EventEmitter;
 
 function runTest(highWaterMark, objectMode, produce) {
