+++ test/parallel/test-stream2-writable.js
@@ -22,8 +22,8 @@
 'use strict';
 
 const common = require('../common');
-const W = require('_stream_writable');
-const D = require('_stream_duplex');
+const W = require('stream').Writable;
+const D = require('stream').Duplex;
 const assert = require('assert');
 
 const util = require('util');
